+++ api/mock/src/commonMain/kotlin/registra/SignUpApiFake.kt
@@ -1,0 +1,7 @@
+package registra
+
+import bitframe.ApiConfigMock
+
+class SignUpApiFake(config: ApiConfigMock = ApiConfigMock()) : SignUpApiDaod(config) {
+    override fun toString(): String = "SignUpApiFake"
+}