+++ core/src/commonMain/kotlin/registra/Product.kt
@@ -1,0 +1,11 @@
+@file:JsExport
+
+package registra
+
+import kotlin.js.JsExport
+
+data class Product(
+    val code: String,
+    val name: String,
+    val logo: String?
+)