+++ core/src/commonMain/kotlin/registra/SignUpRequest.kt
@@ -1,0 +1,7 @@
+package registra
+
+data class SignUpRequest(
+    val name: String,
+    val email: String,
+
+)