+++ rest/src/commonMain/kotlin/registra/RegistraEndpoint.kt
@@ -1,0 +1,5 @@
+package registra
+
+class RegistraEndpoint(val root: String) {
+    val base = "$root/registra"
+}