+++ sdk/server/core/src/commonMain/kotlin/registra/RegistraInstaller.kt
@@ -1,0 +1,7 @@
+package registra
+
+import bitframe.ApplicationConfigBuilder
+
+fun <S> ApplicationConfigBuilder<S>.installRegistra(builder: (S) -> RegistraServer) {
+
+}